Google Analytics: can we do better?
Quite frankly, I'd like to begin this by saying: I'm a frequent user of P2P networks, such as Bittorrent, IPFS and others.

I think P2P is a perfect example of how the people can beat oppression, be it from the state or elsewhere. These technologies are hard to block, and give freedom back to the masses. I personally love the concept behind P2P, and I think we can all say: the concept is taking off quickly, worldwide...

Nevertheless, I would like to suggest this modification to you: remove Google Analytics.

Google Analytics is the antithesis of what P2P tries to achieve, getting around censorship and oppression of the state. In recent times, Google have shown themselves to be at the center of injustice, be it: stealing search history, collecting it and holding it ransom; providing data on BLM activists to the Police. I could go on all day.

Could we work together to remove Google from this website?

Many thanks,
Resynth <resynth1943>
https://resynth1943.net