Fix broken examples link in javascript.md
The current link gives a 404, so I changed it to what I think is the new examples location. As someone new (returning) to libp2p, this is an important link, as it is a natural next step but not easily discovered. 

I also changed the wording slightly since it was an ill-formed sentence and indicated that the source is a part of the main libp2p repo, which is not the case any more.